+++ src/internal/pkg/utils/utils.go
@@ -1,0 +1,10 @@
+package utils
+
+func Contains(slice []string, target string) bool {
+	for _, s := range slice {
+		if s == target {
+			return true
+		}
+	}
+	return false
+}